refactor: card 폴더 순환참조 개선
