+++ src/card/types/CardBrand.type.ts
@@ -1,3 +1,3 @@
-import { CardState } from '@/card';
+import { CardState } from './CardState.type';
 
 export type CardBrand = Pick<CardState, 'label' | 'color'>;
