+++ src/card/utils/isValidateCardBrand.ts
@@ -1,4 +1,4 @@
-import { CardBrandName } from '@/card';
+import { CardBrandName } from '@/card/constants';
 
 export const isValidateCardBrand = (brand: string): brand is CardBrandName =>
   Object.values(CardBrandName).includes(brand as CardBrandName);
