+++ src/card/utils/isValidateCardNumber.ts
@@ -1,4 +1,4 @@
-import type { CardNumber } from '@/card';
+import type { CardNumber } from '@/card/types';
 
 export const isValidateCardNumber = (cardNumber: string): cardNumber is CardNumber =>
   /^[0-9]{4} [0-9]{4} [0-9]{4} [0-9]{4}$/.test(cardNumber);
